requirements.txt
paramiko>=3.4.0
cryptography>=41.0.7
groq>=0.4.0
streamlit>=1.28.0
streamlit-autorefresh>=0.0.1
python-dotenv>=1.0.0
requests>=2.31.0
